fix: return to select auto completion
Summary:
Autocompletion selections with Carriage Returns was broken recently.
this uses a check with the COC Pop Menu and issues the correct command
now.

Test Plan:
Use return on regular non insert mode and confirmed it works. Use return
when popmenu is visible and confirm it selects the highlighted item